Today commit
- 한국 코로나 추가
- 코드 재정리

Error LIST
- CountrySelector 나라 이름이  아닌 숫자로 나오는 현상
