+++ package.json
@@ -1,19 +1,18 @@
-  
 {
-    "name": "wuhan",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-      "start": "next"
-    },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-      "next": "^9.3.0",
-      "normalize.css": "^8.0.1",
-      "react": "^16.13.0",
-      "react-dom": "^16.13.0",
-      "styled-components": "^5.0.1"
-    }
+  "name": "wuhan",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "next"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "next": "^9.3.0",
+    "normalize.css": "^8.0.1",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
+    "styled-components": "^5.0.1"
   }
+}
