+++ pages/index.js
@@ -1,48 +1,13 @@
-import { useState, useEffect } from "react";
-
-function useStats(){
-    const [stats, setStats] = useState(); 
-    useEffect(() => {
-        async function fetchData() {
-        console.log('Fetching Data');
-        const data = await fetch('https://covid19.mathdro.id/api').then(res =>
-         res.json()
-        );
-        setStats(data);
-        }
-    fetchData();
-    }, []);
-    return stats;
-}
-
-function Stats() {
-    const stats = useStats();
-    if(!stats) return <p> loading... </p>
-    //    console.log(stats);
-    return (
-    <div>
-        <div className="statBlock">
-            <h3>Confirmed:</h3>
-            <span>{stats.confirmed.value}</span>
-        </div>
-        
-        <div className="statBlock">
-            <h3>deaths:</h3>
-            <span>{stats.deaths.value}</span>
-        </div>
-
-        <div className="statBlock">
-            <h3>recovered:</h3>
-            <span>{stats.recovered.value}</span>
-        </div>
-    </div>
-    )
-}
+import useStats from '../utils/useStats';
+import Stats from '../components/Stats';
+import CountrySelector from '../components/CountrySelector';
 
 export default function IndexPage() {
     return(
         <div>
-        <Stats></Stats>
+        <Stats url="https://covid19.mathdro.id/api"></Stats>
+        <CountrySelector></CountrySelector>
+        <Stats url="https://covid19.mathdro.id/api/countries/KOR"></Stats>
         </div>
-    )
+    );
 }
